refactor: dropdowns
