+++ apps/app/components/automation/auto-archive-automation.tsx
@@ -28,11 +28,11 @@ export const AutoArchiveAutomation: React.FC<Props> = ({ projectDetails, handleC
         handleClose={() => setmonthModal(false)}
         handleChange={handleChange}
       />
-      <div className="flex flex-col gap-7 px-6 py-5 rounded-[10px] border border-brand-base bg-brand-base">
-        <div className="flex items-center justify-between gap-x-8 gap-y-2 ">
+      <div className="flex flex-col gap-7 px-6 py-5 rounded-[10px] border border-custom-border-100 bg-custom-background-90">
+        <div className="flex items-center justify-between gap-x-8 gap-y-2">
           <div className="flex flex-col gap-2.5">
             <h4 className="text-lg font-semibold">Auto-archive closed issues</h4>
-            <p className="text-sm text-brand-secondary">
+            <p className="text-sm text-custom-text-200">
               Plane will automatically archive issues that have been completed or cancelled for the
               configured time period.
             </p>
@@ -52,17 +52,12 @@ export const AutoArchiveAutomation: React.FC<Props> = ({ projectDetails, handleC
             <div className="w-1/2 text-base font-medium">
               Auto-archive issues that are closed for
             </div>
-            <div className="w-1/2 ">
+            <div className="w-1/2">
               <CustomSelect
                 value={projectDetails?.archive_in}
-                customButton={
-                  <button className="flex w-full items-center justify-between gap-1 rounded-md border border-brand-base shadow-sm duration-300 text-brand-secondary hover:text-brand-base hover:bg-brand-surface-2 focus:outline-none px-3 py-2 text-sm text-left">
-                    {`${projectDetails?.archive_in} ${
-                      projectDetails?.archive_in === 1 ? "Month" : "Months"
-                    }`}
-                    <ChevronDownIcon className="h-3 w-3" aria-hidden="true" />
-                  </button>
-                }
+                label={`${projectDetails?.archive_in} ${
+                  projectDetails?.archive_in === 1 ? "Month" : "Months"
+                }`}
                 onChange={(val: number) => {
                   handleChange({ archive_in: val });
                 }}
