+++ apps/app/components/automation/auto-close-automation.tsx
@@ -77,11 +77,11 @@ export const AutoCloseAutomation: React.FC<Props> = ({ projectDetails, handleCha
         handleChange={handleChange}
       />
 
-      <div className="flex flex-col gap-7 px-6 py-5 rounded-[10px] border border-brand-base bg-brand-base">
+      <div className="flex flex-col gap-7 px-6 py-5 rounded-[10px] border border-custom-border-100 bg-custom-background-90">
         <div className="flex items-center justify-between gap-x-8 gap-y-2 ">
           <div className="flex flex-col gap-2.5">
             <h4 className="text-lg font-semibold">Auto-close inactive issues</h4>
-            <p className="text-sm text-brand-secondary">
+            <p className="text-sm text-custom-text-200">
               Plane will automatically close the issues that have not been updated for the
               configured time period.
             </p>
@@ -102,17 +102,12 @@ export const AutoCloseAutomation: React.FC<Props> = ({ projectDetails, handleCha
               <div className="w-1/2 text-base font-medium">
                 Auto-close issues that are inactive for
               </div>
-              <div className="w-1/2 ">
+              <div className="w-1/2">
                 <CustomSelect
                   value={projectDetails?.close_in}
-                  customButton={
-                    <button className="flex w-full items-center justify-between gap-1 rounded-md border border-brand-base shadow-sm duration-300 text-brand-secondary hover:text-brand-base hover:bg-brand-surface-2 focus:outline-none px-3 py-2 text-sm text-left">
-                      {`${projectDetails?.close_in} ${
-                        projectDetails?.close_in === 1 ? "Month" : "Months"
-                      }`}
-                      <ChevronDownIcon className="h-3 w-3" aria-hidden="true" />
-                    </button>
-                  }
+                  label={`${projectDetails?.close_in} ${
+                    projectDetails?.close_in === 1 ? "Month" : "Months"
+                  }`}
                   onChange={(val: number) => {
                     handleChange({ close_in: val });
                   }}
@@ -143,42 +138,34 @@ export const AutoCloseAutomation: React.FC<Props> = ({ projectDetails, handleCha
                   value={
                     projectDetails?.default_state ? projectDetails?.default_state : defaultState
                   }
-                  customButton={
-                    <button
-                      className={`flex w-full items-center justify-between gap-1 rounded-md border border-brand-base shadow-sm duration-300 text-brand-secondary hover:text-brand-base hover:bg-brand-surface-2 focus:outline-none px-3 py-2 text-sm text-left ${
-                        !multipleOptions ? "opacity-60" : ""
-                      }`}
-                    >
-                      <div className="flex items-center gap-2">
-                        {selectedOption ? (
-                          getStateGroupIcon(selectedOption.group, "16", "16", selectedOption.color)
-                        ) : currentDefaultState ? (
-                          getStateGroupIcon(
-                            currentDefaultState.group,
-                            "16",
-                            "16",
-                            currentDefaultState.color
-                          )
-                        ) : (
-                          <Squares2X2Icon className="h-3.5 w-3.5 text-custom-text-200" />
-                        )}
-                        {selectedOption?.name
-                          ? selectedOption.name
-                          : currentDefaultState?.name ?? (
-                              <span className="text-custom-text-200">State</span>
-                            )}
-                      </div>
-                      {multipleOptions && (
-                        <ChevronDownIcon className="h-3 w-3" aria-hidden="true" />
+                  label={
+                    <div className="flex items-center gap-2">
+                      {selectedOption ? (
+                        getStateGroupIcon(selectedOption.group, "16", "16", selectedOption.color)
+                      ) : currentDefaultState ? (
+                        getStateGroupIcon(
+                          currentDefaultState.group,
+                          "16",
+                          "16",
+                          currentDefaultState.color
+                        )
+                      ) : (
+                        <Squares2X2Icon className="h-3.5 w-3.5 text-custom-text-200" />
                       )}
-                    </button>
+                      {selectedOption?.name
+                        ? selectedOption.name
+                        : currentDefaultState?.name ?? (
+                            <span className="text-custom-text-200">State</span>
+                          )}
+                    </div>
                   }
                   onChange={(val: string) => {
                     handleChange({ default_state: val });
                   }}
                   options={options}
                   disabled={!multipleOptions}
-                  dropdownWidth="w-full"
+                  width="w-full"
+                  input
                 />
               </div>
             </div>
