+++ apps/app/components/issues/view-select/assignee.tsx
@@ -130,7 +130,7 @@ export const ViewAssigneeSelect: React.FC<Props> = ({
       position={position}
       disabled={isNotAllowed}
       selfPositioned={selfPositioned}
-      dropdownWidth="w-full min-w-[12rem]"
+      width="w-full min-w-[12rem]"
     />
   );
 };
