+++ apps/app/components/modules/sidebar-select/select-members.tsx
@@ -70,7 +70,7 @@ export const SidebarMembersSelect: React.FC<Props> = ({ value, onChange }) => {
           }
           options={options}
           onChange={onChange}
-          height="md"
+          maxHeight="md"
           position="right"
           multiple
         />
