+++ apps/app/components/ui/index.ts
@@ -1,16 +1,14 @@
 export * from "./buttons";
+export * from "./dropdowns";
 export * from "./graphs";
 export * from "./input";
 export * from "./text-area";
 export * from "./avatar";
-export * from "./context-menu";
-export * from "./custom-menu";
-export * from "./custom-search-select";
-export * from "./custom-select";
 export * from "./date";
 export * from "./datepicker";
 export * from "./empty-space";
 export * from "./empty-state";
+export * from "./icon";
 export * from "./labels-list";
 export * from "./linear-progress-indicator";
 export * from "./loader";
@@ -25,5 +23,4 @@ export * from "./markdown-to-component";
 export * from "./product-updates-modal";
 export * from "./integration-and-import-export-banner";
 export * from "./range-datepicker";
-export * from "./icon";
 export * from "./circular-progress";
